nicer webpacky
